`gather` and `.` for root.
Implement a `gather` method that will return an array with all the
values for a path along the path.

Use dots univerally, so that `[ "", "" ]` is the root instead of `[ ""
]` so that the root can be expressed as `".".split(".")`. This means
that double dots are invalid because `"...".split(".")` would indicate
the root.

Or maybe every entry should point to itself if you have an empty string?
Then you have rules you're following and you're following them
universally anywhere in the path, which might simplify documentation.